Add Discord Slash Commands

+++ src/plugins/add.js
@@ -1,6 +1,7 @@
 'use strict';
 const ytdl = require('ytdl-core');
 const fs = require('fs');
+const validUrl = require('valid-url');
 const {
   PATH,
   MAX_YT_TIME,
@@ -8,9 +9,102 @@ const {
   ytdownload,
   makeEmbed,
   hasFile,
+  isDiscordCommand,
+  getReplyFunction,
 } = require('./util');
 const USAGE = '`usage: [!add/!a] [<url>/"Example String Here"] <name>' +
   ' [<startTimeStamp>] [<stopTimeStamp>]`';
+
+const downloadAndAdd = ({ fileName, url, exten, stopTime, startTime, loadingMsg, originalMessage }) => {
+  let replyFunction = getReplyFunction(originalMessage);
+  const author = originalMessage?.author || originalMessage?.user;
+  if (!fileName) {
+    if (loadingMsg)
+      loadingMsg.delete();
+    replyFunction(USAGE);
+    return;
+  }
+  // Check for youtube videos
+  if (url.includes('www.youtube.com') || url.includes('youtu.be')) {
+    ytdl.getBasicInfo(url).then((info) => {
+      const duration = (stopTime - startTime) ||
+      (info.videoDetails.lengthSeconds - (startTime || 0));
+      // Check the length of the video
+      if (info.videoDetails.lengthSeconds < MAX_YT_TIME
+        || (duration && duration < MAX_YT_TIME)) {
+        /**
+         * Download the audio from the video as mp3
+         *
+         * Duration is either stopTime - startTime or
+         * the video length - startTime/0
+         */
+        ytdownload({
+          url,
+          fileName,
+          timeStart: startTime,
+          duration,
+          cb: () => {
+            if (loadingMsg)
+              loadingMsg.delete();
+            replyFunction(
+              makeEmbed({
+                message: `Added ${info.videoDetails.title} as ${fileName}`,
+                user: author,
+                member: originalMessage.guild.members.cache.get(author.id).displayName,
+                color: originalMessage.guild.members.cache.get(author.id).displayColor,
+              })
+            );
+          },
+        });
+      } else {
+        if (loadingMsg)
+          loadingMsg.delete();
+        replyFunction(
+          'That video is too long! Keep it at 2:30 or below.'
+        );
+      }
+    }).catch((err) => {
+      console.log('Could not get youtube info', err);
+      if (loadingMsg)
+        loadingMsg.delete();
+      replyFunction('Could not get video info.');
+      return;
+    });
+  } else {
+    if (!exten) {
+      if (loadingMsg)
+        loadingMsg.delete();
+      replyFunction('Could not find file extension');
+    }
+    // Download the file
+    download({
+      url,
+      fileName,
+      extension: exten,
+      timeStart: startTime,
+      timeStop: stopTime,
+      cb: (err) => {
+        if (err) {
+          console.log('Could not download file: ', err);
+          if (loadingMsg)
+            loadingMsg.delete();
+          replyFunction(err);
+        } else {
+          if (loadingMsg)
+            loadingMsg.delete();
+          replyFunction(
+            makeEmbed({
+              message: `Added: ${fileName}`,
+              user: author,
+              member: originalMessage.guild.members.cache.get(author.id).displayName,
+              color: originalMessage.guild.members.cache.get(author.id).displayColor,
+            })
+          );
+        }
+      },
+    });
+  }
+};
 
 /**
  * Adds a file to the local storage space. If the url is a youtube video,
@@ -28,93 +122,36 @@ const USAGE = '`usage: [!add/!a] [<url>/"Example String Here"] <name>' +
  * music in seconds
  */
 const add = (fileName, url, exten, message, startTime, stopTime) => {
+  let replyFunction = getReplyFunction(message);
   if (hasFile({fileName})) {
-    message.channel.send('File name already exists.');
+    replyFunction('File name already exists.');
     return;
   }
   // Send a loading message that will get deleted later since
   // downloading can take a while
-  message.channel.send('Loading...').then(msg => {
-    if (!fileName) {
-      msg.delete();
-      message.channel.send(USAGE);
-      return;
-    }
-    // Check for youtube videos
-    if (url.includes('www.youtube.com') || url.includes('youtu.be')) {
-      ytdl.getBasicInfo(url).then((info) => {
-        const duration = (stopTime - startTime) ||
-        (info.videoDetails.lengthSeconds - (startTime || 0));
-        // Check the length of the video
-        if (info.videoDetails.lengthSeconds < MAX_YT_TIME
-          || (duration && duration < MAX_YT_TIME)) {
-          /**
-           * Download the audio from the video as mp3
-           *
-           * Duration is either stopTime - startTime or
-           * the video length - startTime/0
-           */
-          ytdownload({
-            url,
-            fileName,
-            timeStart: startTime,
-            duration,
-            cb: () => {
-              msg.delete();
-              message.channel.send(
-                makeEmbed({
-                  message: `Added ${info.videoDetails.title} as ${fileName}`,
-                  user: message.author,
-                  member: message.guild.member(message.author.id).displayName,
-                  color: message.guild.member(message.author.id).displayColor,
-                })
-              );
-            },
-          });
-        } else {
-          msg.delete();
-          message.channel.send(
-            'That video is too long! Keep it at 2:30 or below.'
-          );
-        }
-      }).catch((err) => {
-        console.log('Could not get youtube info', err);
-        msg.delete();
-        message.channel.send('Could not get video info.');
-        return;
-      });
-    } else {
-      if (!exten) {
-        msg.delete();
-        message.channel.send('Could not find file extension');
-      }
-      // Download the file
-      download({
-        url,
+  if (isDiscordCommand(message)) {
+    message.defer();
+    downloadAndAdd({
+      fileName,
+      url,
+      exten,
+      originalMessage: message,
+      startTime,
+      stopTime,
+    });
+  } else {
+    message.channel.send('Loading...').then(loadingMsg => {
+      downloadAndAdd({
         fileName,
-        extension: exten,
-        timeStart: startTime,
-        timeStop: stopTime,
-        cb: (err) => {
-          if (err) {
-            console.log('Could not download file: ', err);
-            msg.delete();
-            message.channel.send(err);
-          } else {
-            msg.delete();
-            message.channel.send(
-              makeEmbed({
-                message: `Added: ${fileName}`,
-                user: message.author,
-                member: message.guild.member(message.author.id).displayName,
-                color: message.guild.member(message.author.id).displayColor,
-              })
-            );
-          }
-        },
+        url,
+        exten,
+        originalMessage: message,
+        loadingMsg,
+        startTime,
+        stopTime,
       });
-    }
-  });
+    });
+  }
 };
 
 /**
@@ -126,28 +163,31 @@ const add = (fileName, url, exten, message, startTime, stopTime) => {
  * the command
  */
 const addText = (fileName, text, message) => {
+  let replyFunction = getReplyFunction(message);
+  const author = message?.author || message?.user;
   if (hasFile({fileName})) {
-    message.channel.send('File name already exists.');
+    replyFunction('File name already exists.');
     return;
   }
   fs.writeFile(`${PATH}/${fileName}.txt`, text, (err) => {
     if (err) {
-      message.channel.send('Could not write the text to a file');
+      replyFunction('Could not write the text to a file');
       return;
     }
-    message.delete();
-    message.channel.send(
+    if (!isDiscordCommand(message))
+      message.delete();
+    replyFunction(
       makeEmbed({
         message: `Added: ${fileName}`,
-        user: message.author,
-        member: message.guild.member(message.author.id).displayName,
-        color: message.guild.member(message.author.id).displayColor,
+        user: author,
+        member: message.guild.members.cache.get(author.id).displayName,
+        color: message.guild.members.cache.get(author.id).displayColor,
       })
     );
   });
 };
 
-const onText = (message) => {
+const handleDiscordMessage = (message) => {
   const cmd = message.content.split(' ');
   const botCommand = cmd[0];
   // Get any attachments associated with message
@@ -203,8 +243,68 @@ const onText = (message) => {
   }
 };
 
+const handleDiscordCommand = (interaction) => {
+  if (interaction.commandName === 'add') {
+    const file = interaction.options[0]?.value;
+    const fileName = interaction.options[1]?.value;
+    const timeStart = interaction.options[2]?.value;
+    const timeStop = interaction.options[3]?.value;
+    if (validUrl.isUri(file)) {
+      const exten = file.substr((file.lastIndexOf('.') + 1));
+      if (!exten) {
+        interaction.reply('Could not find file extension');
+      }
+      add(fileName, file, exten, interaction, timeStart, timeStop);
+    } else {
+      addText(fileName, file, interaction);
+    }
+  }
+};
+
+const onText = (discordTrigger) => {
+  if (isDiscordCommand(discordTrigger)) {
+    handleDiscordCommand(discordTrigger);
+  } else {
+    handleDiscordMessage(discordTrigger);
+  }
+};
+
+const commandData = [
+  {
+    name: 'add',
+    description: 'Stores a file to be used with either the play or post command.',
+    options: [
+      {
+        name: 'url_or_text',
+        type: 'STRING',
+        description: 'The file you want to add. Can either be a url (such as a youtube or image link) or plain text.',
+        required: true,
+      },
+      {
+        name: 'name',
+        type: 'STRING',
+        description: 'The name you want the file to be stored as.',
+        required: true,
+      },
+      {
+        name: 'time_start',
+        type: 'INTEGER',
+        description: 'If uploading a youtube video, this signifies the start time in seconds.',
+        required: false,
+      },
+      {
+        name: 'time_end',
+        type: 'INTEGER',
+        description: 'If uploading a youtube video, this signifies the end time in seconds.',
+        required: false,
+      }
+    ],
+  }
+];
+
 module.exports = {
   add,
   addText,
   onText,
+  commandData,
 };
